// server/src/domain/enums/routes.enum.ts
export enum AdminRoutes {
  USERS = '/users',
  USERS_BLOCK = '/users/block',
  USERS_ID = '/users/:id',
  COMPANIES = '/companies',
  COMPANIES_VERIFICATION = '/companies/verification',
  COMPANIES_VERIFY = '/companies/verify',
  JOBS = '/jobs',
  JOBS_STATS = '/jobs/stats',
  JOBS_ID = '/jobs/:id',
  JOBS_ID_STATUS = '/jobs/:id/status',
  JOB_CATEGORIES = '/job-categories',
  JOB_CATEGORIES_ID = '/job-categories/:id',
  SKILLS = '/skills',
  SKILLS_ID = '/skills/:id',
  JOB_ROLES = '/job-roles',
  JOB_ROLES_ID = '/job-roles/:id',
  SUBSCRIPTION_PLANS = '/subscription-plans',
  SUBSCRIPTION_PLANS_ID = '/subscription-plans/:id',
  SUBSCRIPTION_PLANS_MIGRATE = '/subscription-plans/:id/migrate-subscribers',
  PAYMENT_ORDERS = '/payment-orders',
}